+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.0.1657-alpha]- 2025-07-02
+
+1. Minmax, max, min, clamp and calc support.
+2. Mixed units support (rem -> vw)
+3. Performance adjustments
+
 ## [0.0.1656-alpha]- 2025-06-30
 
 ### Fixed